improved pageSkip heuristic
skips will require twice the time before being initiated if they were
needed rarely so far (less than 30% of the scanned pages)  or if the
last page was already skipped.

this slows down the scanning of the first few pages of fixed format
mangas, which is acceptable because of the noticably increased backup
consistency for all comic types.