new page skip implementation
the dual-page comic decision is now made after a certain threshold of compatible indicators was collected.

if a comic is non dual-page according to that definition, it will now be made sure, that every page gets included by reopening missed pages until they could be scanned (this happens because it is not possible to safely determine if a certain page is actually opened in the reader).